Move transfer agreement resolvers into dedicated modules
